pyproject.toml
[project]
name = "ai-book-writer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "mfahadjbr", email = "mfahadjbr@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.100.1",
    "crewai-tools>=0.33.0",
    "pdfkit>=1.0.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
    "toml>=0.10.2",
    "tornado>=6.4.2",
    "watchdog>=6.0.0",
]

[project.scripts]
ai-book-writer = "ai_book_writer:main"
fahad="ai_book_writer.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
